+++ src/main/java/wooteco/subway/exception/station/StationDuplicationException.java
@@ -1,6 +1,7 @@
-package wooteco.subway.exception;
+package wooteco.subway.exception.station;
 
 import org.springframework.http.HttpStatus;
+import wooteco.subway.exception.SubwayException;
 
 public class StationDuplicationException extends SubwayException {
     public StationDuplicationException() {
