+++ src/main/java/wooteco/subway/line/service/LineService.java
@@ -1,8 +1,11 @@
-package wooteco.subway.line;
+package wooteco.subway.line.service;
 
 import org.springframework.stereotype.Service;
-import wooteco.subway.exception.LineDuplicationException;
-import wooteco.subway.exception.NoLineException;
+import wooteco.subway.exception.line.LineDuplicationException;
+import wooteco.subway.exception.line.NoLineException;
+import wooteco.subway.line.domain.Line;
+import wooteco.subway.line.dao.LineDao;
+import wooteco.subway.line.dto.LineRequest;
 
 import java.util.List;
