+++ src/main/java/wooteco/subway/station/dao/StationDao.java
@@ -1,4 +1,6 @@
-package wooteco.subway.station;
+package wooteco.subway.station.dao;
+
+import wooteco.subway.station.domain.Station;
 
 import java.util.List;
 import java.util.Optional;
