+++ src/main/java/wooteco/subway/station/dao/StationH2Dao.java
@@ -1,10 +1,10 @@
-package wooteco.subway.station;
+package wooteco.subway.station.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import wooteco.subway.line.Line;
+import wooteco.subway.station.domain.Station;
 
 import java.sql.PreparedStatement;
 import java.util.List;
