+++ src/test/java/wooteco/subway/line/LineAcceptanceTest.java
@@ -9,13 +9,12 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.jdbc.Sql;
 import wooteco.subway.AcceptanceTest;
-import wooteco.subway.exception.LineDuplicationException;
-import wooteco.subway.exception.StationDuplicationException;
+import wooteco.subway.exception.line.LineDuplicationException;
+import wooteco.subway.line.dto.LineRequest;
+import wooteco.subway.line.dto.LineResponse;
 
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
