+++ src/test/java/wooteco/subway/line/LineH2DaoTest.java
@@ -6,6 +6,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
+import wooteco.subway.line.dao.LineDao;
+import wooteco.subway.line.domain.Line;
 
 import java.util.Arrays;
 import java.util.List;
